add installPage to NodeClient
